+++ components/Foot.tsx
@@ -20,7 +20,6 @@ const styles = Native.StyleSheet.create({
         paddingRight: 40,
         paddingTop: 30,
         paddingBottom: 30,
-        // NOTA: No sé por qué, pero VSCode me marca aquí un error. Eso no pasaba con JS. Además, ahora ya no funciona (sin TS si lo hacía). Extraño...
         position: "absolute",
         zIndex: 99,
         bottom: 0,